backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2docx
python-docx
